rename k8s package name
