+++ go/controller/k8s/api.go
@@ -12,7 +12,7 @@
    See the License for the specific language governing permissions and
 	 limitations under the License. */
 
-package controller
+package k8s
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
